refactor: put renderSVG logic in separate folder
